Implemented pending changes adding. Content setters can now be called from non-main threads.

+++ Apps/ArtCode/CodeFile+Generation.h
@@ -8,7 +8,7 @@
 
 #import "CodeFile.h"
 
-typedef int32_t CodeFileGeneration;
+typedef NSUInteger CodeFileGeneration;
 
 @interface CodeFile (Generation)
